Complete recoding

+++ src/main/java/org/drachens/dataClasses/Diplomacy/Justifications/WarGoalType.java
@@ -2,5 +2,5 @@ package org.drachens.dataClasses.Diplomacy.Justifications;
 
 import org.drachens.dataClasses.additional.Modifier;
 
-public record WarGoalType(String name, Modifier modifier, float expires, float timeToMake) {
+public record WarGoalType(String name, Modifier modifier, int expires, int timeToMake) {
 }
